[kotlin2cpg] Some cleanup and refactoring
