+++ joern-cli/frontends/kotlin2cpg/src/main/scala/io/joern/kotlin2cpg/passes/AstCreationPass.scala
@@ -2,7 +2,6 @@ package io.joern.kotlin2cpg.passes
 
 import io.joern.kotlin2cpg.KtFileWithMeta
 import io.joern.kotlin2cpg.ast.AstCreator
-import io.joern.kotlin2cpg.types.{NameRenderer, TypeInfoProvider}
 import io.joern.x2cpg.ValidationMode
 import io.joern.x2cpg.datastructures.Global
 import io.shiftleft.codepropertygraph.generated.Cpg
@@ -12,13 +11,9 @@ import org.slf4j.LoggerFactory
 
 import scala.jdk.CollectionConverters.EnumerationHasAsScala
 
-class AstCreationPass(
-  filesWithMeta: Iterable[KtFileWithMeta],
-  typeInfoProvider: TypeInfoProvider,
-  bindingContext: BindingContext,
-  cpg: Cpg
-)(implicit withSchemaValidation: ValidationMode)
-    extends ForkJoinParallelCpgPass[KtFileWithMeta](cpg) {
+class AstCreationPass(filesWithMeta: Iterable[KtFileWithMeta], bindingContext: BindingContext, cpg: Cpg)(implicit
+  withSchemaValidation: ValidationMode
+) extends ForkJoinParallelCpgPass[KtFileWithMeta](cpg) {
 
   private val logger         = LoggerFactory.getLogger(getClass)
   private val global: Global = new Global()
@@ -28,7 +23,7 @@ class AstCreationPass(
   override def generateParts(): Array[KtFileWithMeta] = filesWithMeta.toArray
 
   override def runOnPart(diffGraph: DiffGraphBuilder, fileWithMeta: KtFileWithMeta): Unit = {
-    diffGraph.absorb(new AstCreator(fileWithMeta, typeInfoProvider, bindingContext, global).createAst())
+    diffGraph.absorb(new AstCreator(fileWithMeta, bindingContext, global).createAst())
     logger.debug(s"AST created for file at `${fileWithMeta.f.getVirtualFilePath}`.")
   }
